requirements.txt
deepface
firebase_admin
flask
flask_cors
numpy
Pillow
tf-keras